Show custom error page if origin responds with error
This is mostly done because it was found by VMon that when
the origin sends a non standard HTTP error message then
the traffic server halts for approx one minute. If we handle
such errors ourselves the response to the client is immediate.

See https://redmine.equalit.ie/issues/2089 for more info.