Fix race condition happening on config reloads (hopefully)
Instead of modifying the Banjax global object when a management callback
is executed (while there may be other transactions still running). Every
transaction keeps a const reference to a shared Banjax instance. Then,
when the namagement callback is executed, we create a new Banjax
instance which will be used for consecutive transactions.

Kudos to @redlizard for suggesting this approach.

Co-Authored-By: Ruud Koolen <redlizard@redlizard.nl>
Co-Authored-By: Peter Jankuliak <p.jankuliak@gmail.com>